+++ arinc424/src/fields/unit_ind.rs
@@ -1,0 +1,39 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Joe Pearson
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::{Error, FixedField};
+
+/// 5.133 Unit Indicator (UNIT IND)
+#[derive(Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
+pub enum UnitIndicator {
+    MeanSeaLevel,
+    AboveGroundLevel,
+}
+
+impl FixedField<'_> for UnitIndicator {
+    const LENGTH: usize = 1;
+
+    fn from_bytes(bytes: &'_ [u8]) -> Result<Self, Error> {
+        match bytes[0] {
+            b'M' => Ok(Self::MeanSeaLevel),
+            b'A' => Ok(Self::AboveGroundLevel),
+            byte => Err(crate::Error::InvalidCharacter {
+                field: "Unit Indicator",
+                byte,
+                expected: "M or A",
+            }),
+        }
+    }
+}